requirements.txt
aiofiles==24.1.0
aiogram==3.17.0
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
async-timeout==5.0.1
attrs==24.3.0
certifi==2024.12.14
charset-normalizer==3.4.1
distro==1.9.0
exceptiongroup==1.2.2
frozenlist==1.5.0
googletrans==4.0.2
groq==0.13.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.7
httpx==0.28.1
hyperframe==6.0.1
idna==3.10
magic-filter==1.0.12
multidict==6.1.0
propcache==0.2.1
pydantic==2.10.4
pydantic_core==2.27.2
python-dotenv==1.0.1
requests==2.32.3
requests-toolbelt==1.0.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
pymongo~=4.10.1